// src/models/LibraryError.ts
export enum ErrorCode {
  BOOK_NOT_FOUND = 'BOOK_NOT_FOUND',
  BOOK_ALREADY_BORROWED = 'BOOK_ALREADY_BORROWED',
  BOOK_NOT_BORROWED = 'BOOK_NOT_BORROWED',
  EMPLOYEE_NOT_FOUND = 'EMPLOYEE_NOT_FOUND',
  LOAN_LIMIT_EXCEEDED = 'LOAN_LIMIT_EXCEEDED',
  CANNOT_DELETE_BORROWED_BOOK = 'CANNOT_DELETE_BORROWED_BOOK',
  CANNOT_DELETE_EMPLOYEE_WITH_LOANS = 'CANNOT_DELETE_EMPLOYEE_WITH_LOANS',
  DATA_SAVE_FAILED = 'DATA_SAVE_FAILED',
  DATA_LOAD_FAILED = 'DATA_LOAD_FAILED',
  INVALID_INPUT = 'INVALID_INPUT',
  // 書籍情報取得関連
  BOOK_INFO_NOT_FOUND = 'BOOK_INFO_NOT_FOUND',
  BOOK_INFO_FETCH_FAILED = 'BOOK_INFO_FETCH_FAILED',
  BOOK_INFO_TIMEOUT = 'BOOK_INFO_TIMEOUT',
  BOOK_INFO_NETWORK_ERROR = 'BOOK_INFO_NETWORK_ERROR',
}

export class LibraryError extends Error {
  constructor(
    message: string,
    public code: ErrorCode
  ) {
    super(message);
    this.name = 'LibraryError';
  }
}